Send original manifest dictionary
